requirements.txt
Flask==2.1.1
PyJWT==2.3.0
mysql-connector-python==8.0.27